Add version command and decouple evidence requests
