+++ turbinia/api/cli/turbinia_client/core/groups.py
@@ -15,6 +15,13 @@
 """Turbinia API client command-line tool."""
 
 import click
+import logging
+import sys
+
+from turbinia_api_lib import exceptions
+from turbinia_client.factory import factory
+
+log = logging.getLogger('turbinia')
 
 
 @click.group('config')
@@ -42,6 +49,32 @@ def jobs_group():
   """Get a list of enabled Turbinia jobs."""
 
 
-@click.group('submit')
-def submit_group():
-  """Submit new requests to the Turbinia API server."""
+@click.pass_context
+def setup_submit(ctx: click.Context):
+  try:
+    ctx.obj.evidence_mapping = ctx.obj.get_evidence_arguments()
+    ctx.obj.request_options = ctx.obj.get_request_options()
+
+    # Build all the commands based on responses from the API server.
+    request_commands = factory.CommandFactory.create_dynamic_objects(
+        evidence_mapping=ctx.obj.evidence_mapping,
+        request_options=ctx.obj.request_options)
+    for command in request_commands:
+      submit_group.add_command(command)
+  except exceptions.ApiException as exception:
+    log.error(
+        'Error while attempting to contact the API server during setup: %s',
+        exception)
+    sys.exit(-1)
+
+
+@click.group('submit', chain=True, invoke_without_command=True)
+@click.pass_context
+def submit_group(ctx: click.Context):
+  """Submit new requests to the Turbinia API server.
+  
+  Please run this command without any arguments to view a list
+  of available evidence types.
+  """
+  ctx.invoke(setup_submit)
+  click.echo(submit_group.get_help(ctx))
